+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 18/07/2022 - v3.5.5
+
+- **File formats:**
+   - (Add) `LayerImageType`: Gets the layer image data type used on this file format
+   - (Improvement) jxs, rgb.cws and xml.cws: Improve the layer image read/write performance by a significant amount
+   - (Fix) xml.cws: Wanhao printers need 32 bit png instead of 8 bit png (#514)
+
 ## 14/07/2022 - v3.5.4
 
 - **PCB Exposure:**